Former-commit-id: 820f76b39158acac0e3627b4616af736d4d86d2e

+++ BE/arena/src/main/java/com/example/arena/domain/community/dto/response/SearchBoardResponse.java
@@ -12,6 +12,6 @@ public class SearchBoardResponse {
 	private String title;
 	private String nickname;
 //	private Date createdAt; 구현 필요
-	private int comments;
+	private long comments;
 	private int likes;
 }
